+++ NEWS.md
@@ -1,9 +1,14 @@
-# PLNmodels dev version 
+# PLNmodels 1.0.4 (2023-08-24)
 
 * changed initial value in optim for variational variance (1 -> 0.1) in VE-step of PLN and PLNPCA
 * fix sign in objective of VE_step for PLN with full covariance Issue #100
+* add a `scale` argument compute_offset() to force the offsets (RLE, CSS, GMPR, Wrench) to be on the same scale as the counts, like TSS.
+* add a new "TMM" for compute_offset()
+* fix nb_param for PLNLDA, which caused wrong BIC/ICL and erratic model selection
+* fix minor issues #102, #103 plus some others
+* fix package file documentation as suggested in <https://github.com/r-lib/roxygen2/issues/1491>
 
-# PLNmodels 1.0.3 
+# PLNmodels 1.0.3 (2023-07-06)
 
 * higher tolerance on a single test (among 700) that fails on the 'noLD' 
   additional architecture on CRAN (tests without long double)